requirements.txt
Flask
pymongo

requests
oauth2client
google-api-python-client
msal
fastapi
uvicorn
numpy
srsly
starlette
orjson
schedule
pytz
anyio
pillow